+++ main/cpp/gengeopop/generators/K12SchoolGenerator.cpp
@@ -35,7 +35,8 @@ void K12SchoolGenerator::Apply(shared_ptr<GeoGrid> geoGrid, const GeoGridConfig&
         //    to the relative population w.r.t the total population.
 
         const auto pupilCount  = geoGridConfig.popInfo.popcount_k12school;
-        const auto schoolCount = static_cast<unsigned int>(ceil(pupilCount / geoGridConfig.pools.k12school_size));
+        const auto schoolCount = static_cast<unsigned int>(ceil(pupilCount
+                / static_cast<double>(geoGridConfig.pools.k12school_size)));
 
         vector<double> weights;
         for (const auto& loc : *geoGrid) {
